+++ apps/prairielearn/src/ee/pages/administratorInstitutionCourses/administratorInstitutionCourses.html.ts
@@ -27,9 +27,12 @@ export function AdministratorInstitutionCourses({
           navPage: 'administrator_institution',
           navSubPage: 'courses',
         })}
-        <main class="container mb-4">
+        <main id="content" class="container mb-4">
           <div class="table-responsive">
-            <table class="table table-hover table-striped table-bordered table">
+            <table
+              class="table table-hover table-striped table-bordered table"
+              aria-label="Courses"
+            >
               <thead>
                 <tr>
                   <th>Name</th>