+++ apps/prairielearn/src/ee/pages/administratorInstitutionLti13/administratorInstitutionLti13.html.ts
@@ -38,7 +38,7 @@ export function AdministratorInstitutionLti13({
           navPage: 'administrator_institution',
           navSubPage: 'lti13',
         })}
-        <main class="container mb-4">
+        <main id="content" class="container mb-4">
           <h2 class="h4">LTI 1.3 / Learning Tools Interoperability</h2>
           <p>
             ${lti13Instances.length} instance${lti13Instances.length === 1 ? '' : 's'} configured.
@@ -233,7 +233,13 @@ ${JSON.stringify(instance.issuer_params, null, 3)}</textarea
 ${JSON.stringify(instance.custom_fields, null, 3)}</textarea
           >
           <small id="custom_fieldsHelp" class="form-text text-muted">
-            Provide suggestions to the LMS in the config JSON for how to setup LTI 1.3 custom fields
+            Provide suggestions to the LMS in the config JSON for how to setup LTI 1.3 custom
+            fields.
+            <a
+              href="https://canvas.instructure.com/doc/api/file.tools_variable_substitutions.html"
+              target="_blank"
+              >Canvas variable substitution docs</a
+            >
           </small>
         </div>
 
@@ -337,7 +343,9 @@ ${JSON.stringify(instance.custom_fields, null, 3)}</textarea
           <small id="uinAttributeHelp" class="form-text text-muted">
             The UIN is used as an internal, immutable identifier for the user. It
             <strong>MUST</strong> never change for a given individual, even if they change their
-            name or UID.
+            name or UID. Possibly
+            <code>["https://purl.imsglobal.org/spec/lti/claim/lis"]["person_sourcedid"]</code> or
+            <code>["https://purl.imsglobal.org/spec/lti/claim/custom"]["uin"]</code>
           </small>
         </div>