+++ apps/prairielearn/src/ee/pages/instructorInstanceAdminLti13/instructorInstanceAdminLti13.html.ts
@@ -39,9 +39,11 @@ export function InstructorInstanceAdminLti13({
           ...resLocals,
           navSubPage: 'lti13',
         })}
-        <main class="container-fluid mb-4">
+        <main id="content" class="container-fluid mb-4">
           <div class="card mb-4">
-            <div class="card-header bg-primary text-white d-flex">LTI 1.3 configuration</div>
+            <div class="card-header bg-primary text-white d-flex">
+              <h1>LTI 1.3 configuration</h1>
+            </div>
             <div class="card-body">
               <div class="row">
                 <div class="col-2">