+++ apps/prairielearn/src/pages/authPassword/authPassword.html.ts
@@ -21,7 +21,8 @@ export function AuthPassword({
           ...resLocals,
           navPage: '',
         })}
-        <main class="container">
+        <main id="content" class="container">
+          <h1 class="sr-only">Assessment Password</h1>
           <div class="card mb-4">
             <div class="card-body">
               <p class="text-center">A password is required to access this assessment.</p>