+++ apps/prairielearn/src/pages/instructorQuestionStatistics/instructorQuestionStatistics.html.ts
@@ -61,11 +61,14 @@ export function InstructorQuestionStatistics({
 
           <div class="card mb-4">
             <div class="card-header bg-primary text-white">
-              Detailed assessment statistics for question ${resLocals.question.qid}
+              <h1>Detailed assessment statistics for question ${resLocals.question.qid}</h1>
             </div>
 
             <div class="table-responsive">
-              <table class="table table-sm table-hover tablesorter table-bordered">
+              <table
+                class="table table-sm table-hover tablesorter table-bordered"
+                aria-label="Question statistics by assessment"
+              >
                 <thead>
                   <tr>
                     <th class="text-center">Course Instance</th>