+++ apps/prairielearn/src/pages/studentAssessmentInstanceFile/studentAssessmentInstanceFile.js
@@ -2,35 +2,40 @@
 import { Router } from 'express';
 import asyncHandler from 'express-async-handler';
 
-import * as error from '@prairielearn/error';
-import * as sqldb from '@prairielearn/postgres';
+import { HttpStatusError } from '@prairielearn/error';
+import { loadSqlEquiv, queryOptionalRow } from '@prairielearn/postgres';
 
+import { FileSchema } from '../../lib/db-types.js';
 import * as fileStore from '../../lib/file-store.js';
 
-const sql = sqldb.loadSqlEquiv(import.meta.url);
+const sql = loadSqlEquiv(import.meta.url);
 const router = Router();
 
 router.get(
   '/:unsafe_file_id(\\d+)/:unsafe_display_filename',
   asyncHandler(async (req, res, next) => {
-    const params = {
-      assessment_instance_id: res.locals.assessment_instance.id,
-      unsafe_file_id: req.params.unsafe_file_id,
-      unsafe_display_filename: req.params.unsafe_display_filename,
-    };
-
     // Assert that the file belongs to this assessment, that the display
     // filename matches, and that the file is not deleted.
-    const result = await sqldb.queryZeroOrOneRowAsync(sql.select_file, params);
-    if (result.rows.length === 0) {
-      throw new error.HttpStatusError(404, 'File not found');
+    const file = await queryOptionalRow(
+      sql.select_file,
+      {
+        assessment_instance_id: res.locals.assessment_instance.id,
+        unsafe_file_id: req.params.unsafe_file_id,
+        unsafe_display_filename: req.params.unsafe_display_filename,
+      },
+      FileSchema,
+    );
+    if (file == null) {
+      throw new HttpStatusError(404, 'File not found');
+    }
+    if (file.type === 'student_upload' && !res.locals.assessment.allow_personal_notes) {
+      throw new HttpStatusError(403, 'Assessment does not allow access to personal notes');
     }
 
-    const { id: fileId, display_filename: displayFilename } = result.rows[0];
-    const stream = await fileStore.getStream(fileId);
+    const stream = await fileStore.getStream(file.id);
     // Ensure the response is interpreted as an "attachment" (file to be downloaded)
     // and not as a webpage.
-    res.attachment(displayFilename);
+    res.attachment(file.display_filename);
     stream.on('error', next).pipe(res);
   }),
 );