+++ apps/prairielearn/src/server.js
@@ -166,7 +166,7 @@ export async function initExpress() {
       sessionTtl <
       (config.sessionStoreExpireSeconds - config.sessionStoreAutoExtendThrottleSeconds) * 1000
     ) {
-      req.session.setExpiration(config.sessionStoreExpireSeconds);
+      req.session.setExpiration(config.sessionStoreExpireSeconds * 1000);
     }
 
     next();
@@ -1290,6 +1290,13 @@ export async function initExpress() {
     },
     (await import('./pages/instructorQuestions/instructorQuestions.js')).default,
   ]);
+  app.use('/pl/course_instance/:course_instance_id(\\d+)/instructor/ai_generate_question_jobs', [
+    (await import('./ee/pages/instructorAiGenerateJobs/instructorAiGenerateJobs.js')).default,
+  ]);
+  app.use(
+    '/pl/course_instance/:course_instance_id(\\d+)/instructor/ai_generate_question_job/:job_sequence_id(\\d+)',
+    [(await import('./ee/pages/instructorAiGenerateJob/instructorAiGenerateJob.js')).default],
+  );
   app.use('/pl/course_instance/:course_instance_id(\\d+)/instructor/ai_generate_question', [
     function (req, res, next) {
       res.locals.navSubPage = 'questions';
@@ -1543,8 +1550,6 @@ export async function initExpress() {
     (await import('./middlewares/logPageView.js')).default('studentAssessments'),
     (await import('./pages/studentAssessments/studentAssessments.js')).default,
   ]);
-  // Exam/Homeworks student routes are polymorphic - they have multiple handlers, each of
-  // which checks the assessment type and calls next() if it's not the right type
   app.use('/pl/course_instance/:course_instance_id(\\d+)/assessment/:assessment_id(\\d+)', [
     (await import('./middlewares/selectAndAuthzAssessment.js')).default,
     (await import('./middlewares/studentAssessmentAccess.js')).default,
@@ -1833,6 +1838,12 @@ export async function initExpress() {
       next();
     },
     (await import('./pages/instructorQuestions/instructorQuestions.js')).default,
+  ]);
+  app.use('/pl/course/:course_id(\\d+)/ai_generate_question_jobs', [
+    (await import('./ee/pages/instructorAiGenerateJobs/instructorAiGenerateJobs.js')).default,
+  ]);
+  app.use('/pl/course/:course_id(\\d+)/ai_generate_question_job/:job_sequence_id(\\d+)', [
+    (await import('./ee/pages/instructorAiGenerateJob/instructorAiGenerateJob.js')).default,
   ]);
   app.use('/pl/course/:course_id(\\d+)/ai_generate_question', [
     function (req, res, next) {
@@ -2319,6 +2330,14 @@ if (esMain(import.meta) && config.startServer) {
         await opentelemetry.init({
           ...config,
           serviceName: 'prairielearn',
+          // For Sentry to work correctly, it needs to hook into our OpenTelemetry setup.
+          // https://docs.sentry.io/platforms/javascript/guides/node/tracing/instrumentation/opentelemetry/
+          //
+          // However, despite what their documentation claims, only the `SentryContextManager`
+          // is necessary if one isn't using Sentry for tracing. In fact, if `SentrySpanProcessor`
+          // is used, 100% of traces will be sent to Sentry, despite us never having set
+          // `tracesSampleRate` in the Sentry configuration.
+          contextManager: config.sentryDsn ? new Sentry.SentryContextManager() : undefined,
         });
 
         // Same with Sentry configuration.
@@ -2326,6 +2345,23 @@ if (esMain(import.meta) && config.startServer) {
           await Sentry.init({
             dsn: config.sentryDsn,
             environment: config.sentryEnvironment,
+
+            // Sentry (specifically `import-in-the-middle`, which Sentry uses)
+            // is known to cause issues with loading `openai` as ESM. Their
+            // recommended workaround it to exclude the module from their hooks.
+            // See related issues:
+            // https://github.com/openai/openai-node/issues/903
+            // https://github.com/getsentry/sentry-javascript/issues/12414
+            registerEsmLoaderHooks: {
+              exclude: [/openai/],
+            },
+
+            // We have our own OpenTelemetry setup, so ensure Sentry doesn't
+            // try to set that up for itself, but only if OpenTelemetry is
+            // enabled. Otherwise, allow Sentry to install its own stuff so
+            // that request isolation works correctly.
+            skipOpenTelemetrySetup: config.openTelemetryEnabled,
+
             beforeSend: (event) => {
               // This will be necessary until we can consume the following change:
               // https://github.com/chimurai/http-proxy-middleware/pull/823