+++ packages/sentry/src/index.ts
@@ -63,8 +63,16 @@ export function requestHandler() {
   return (req: any, _res: any, next: any) => {
     Sentry.withIsolationScope((scope) => {
       scope.addEventProcessor((event) => {
-        event.transaction = extractTransaction(req);
-        return Sentry.addRequestDataToEvent(event, req);
+        // If an event processor throws an error, Sentry will catch it and
+        // retrigger the event processor, which infinitely recurses. We'll
+        // treat our event processor as a best-effort operation and silently
+        // swallow any errors.
+        try {
+          event.transaction = extractTransaction(req);
+          return Sentry.addRequestDataToEvent(event, req);
+        } catch {
+          return event;
+        }
       });
 
       next();
@@ -78,52 +86,54 @@ export function requestHandler() {
 export {
   Breadcrumb,
   BreadcrumbHint,
-  Request,
-  PolymorphicRequest,
-  SdkInfo,
   Event,
   EventHint,
   Exception,
+  NodeOptions,
+  PolymorphicRequest,
+  Request,
+  SdkInfo,
   Session,
   SeverityLevel,
+  Span,
   StackFrame,
   Stacktrace,
   Thread,
   User,
-  Span,
-  NodeOptions,
 } from '@sentry/node';
 
 export {
-  addEventProcessor,
   addBreadcrumb,
-  captureException,
+  addEventProcessor,
+  addRequestDataToEvent,
   captureEvent,
+  captureException,
   captureMessage,
+  close,
   createTransport,
+  defaultStackParser,
+  expressErrorHandler,
+  expressIntegration,
+  extractRequestData,
+  flush,
   getCurrentHub,
+  getCurrentScope,
+  getSentryRelease,
+  makeNodeTransport,
+  NodeClient,
   Scope,
   SDK_VERSION,
+  SentryContextManager,
   setContext,
   setExtra,
   setExtras,
   setTag,
   setTags,
+  setupExpressErrorHandler,
   setUser,
-  withScope,
-  NodeClient,
-  makeNodeTransport,
-  addRequestDataToEvent,
-  extractRequestData,
-  defaultStackParser,
-  flush,
-  close,
-  getSentryRelease,
-  getCurrentScope,
+  startInactiveSpan,
   startSpan,
   startSpanManual,
-  startInactiveSpan,
-  expressIntegration,
-  expressErrorHandler,
-  setupExpressErrorHandler,
+  withIsolationScope,
+  withScope,
 } from '@sentry/node';