+++ packages/named-locks/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @prairielearn/named-locks
 
+## 3.0.7
+
+### Patch Changes
+
+- Updated dependencies [852c2e2]
+  - @prairielearn/postgres@2.1.3
+
+## 3.0.6
+
+### Patch Changes
+
+- a8438ff: Upgrade all JavaScript dependencies
+- Updated dependencies [a8438ff]
+  - @prairielearn/postgres@2.1.2
+
 ## 3.0.5
 
 ### Patch Changes